LPS-71164 Move ElementComparator
