+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/checks/XMLFriendlyURLRoutesFileCheck.java
@@ -19,8 +19,8 @@ import com.liferay.portal.kernel.util.ReleaseInfo;
 import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Tuple;
-import com.liferay.source.formatter.ElementComparator;
 import com.liferay.source.formatter.SourceFormatterMessage;
+import com.liferay.source.formatter.checks.comparator.ElementComparator;
 import com.liferay.source.formatter.checks.util.SourceUtil;
 
 import java.util.HashSet;
